// src/app/_core/models/person.ts
export interface Person {
  idPerson: number;
  idCustomer: number;
  roleLabel: string;
  roleKey: string;
  typePerson: string;
  firstName: string;
  lastName: string;
  title: string;
  nicheIdentifier: string;
  crmName: string;
  siret: string;
  category: string;
  companyName: string;
  companyNameParentId;
  personRoleLabel: string;
}